Switch to rest-client instead of rest_client (subtle...)
So it turns out the rest_client is the same as rest-client but with
mime-type support removed. At first, my understanding was that they
had switched to underscore for some reason because rest_client was
of a newer version than rest-client. But actually the author of this
gem branched the feature set and only indicated this with an underscore.